requirements.txt
# 核心框架
tensorflow>=2.8.0
scikit-learn>=1.0.0
numpy>=1.21.0
scipy>=1.7.0

# 数据处理和可视化
pandas>=1.3.0
matplotlib>=3.5.0

# 工具库
joblib>=1.1.0
python-dateutil>=2.8.0
tqdm>=4.62.0